+++ packages/scroller/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @smoovy/scroller
 
+## 1.0.7
+
+### Patch Changes
+
+- observer: added scrollSize
+- Updated dependencies
+  - @smoovy/observer@1.0.5
+
 ## 1.0.6
 
 ### Patch Changes